Refactor(soonki/activities): Activities관련 scss 파일 분리를 통한 폴더 구조 개선
