+++ src/components/ClubHome/Activities/ModalHeader.js
@@ -1,5 +1,5 @@
 import React from "react";
-import styles from "../../../styles/Club/Home/Activities/Modal.module.scss";
+import styles from "../../../styles/Club/Home/Activities/ModalHeader.module.scss";
 
 const ModalHeader = ({ desc }) => {
   return (
