+++ src/components/ClubHome/Activities/ModalImg.js
@@ -1,5 +1,5 @@
 import React from "react";
-import styles from "../../../styles/Club/Home/Activities/Modal.module.scss";
+import styles from "../../../styles/Club/Home/Activities/ModalImg.module.scss";
 import { IoIosArrowBack, IoIosArrowForward } from "react-icons/io";
 const ModalImg = ({ onClose, imgURL, back, foward, iconSize }) => {
   return (
