+++ src/styles/Club/Home/Activities/ModalFooter.module.scss
@@ -1,0 +1,17 @@
+.footer {
+  padding: 2vh 0;
+  display: flex;
+  margin-top: 1vh;
+  font-weight: bold;
+  span {
+    font-size: 120%;
+    margin: 0 20vw;
+  }
+}
+
+@media (max-width: 767px) {
+  .footer {
+    position: fixed;
+    top: 80vh;
+  }
+}
