+++ src/styles/Club/Home/Activities/ModalHeader.module.scss
@@ -1,0 +1,10 @@
+.header {
+  display: flex;
+  width: 50vw;
+  margin-bottom: 3vh;
+  justify-content: space-between;
+  span {
+    font-family: "twayair";
+    font-size: 150%;
+  }
+}
